(FM-7014) panos_commit type
This wraps up the work to make the PA systems commit. The plan here is to
have a single `panos_commit` resource, that's defined once that every other
type can `autobefore`, which will cause a single commit happen at the end
of the catalog run.

See the `examples/test_commit.pp` file for example usage.