(FM-7153) puppet-string/tasks added to generate REFERENCE.md
As part of documentation and release process `puppet strings generate --format markdown` has been applied to generate a reference of the `types` and `tasks` available in the PANOS module.

`puppet-string/tasks` added as a rake task and progress added to `Cutting a release` section in README to be completed later.