(maint) updating the ssl_fingerprint to accept spaced and colon SHA256 notation
Previously the fingerprint expected it to be the format `XX:XX:...` althought some browsers would display the fingerprint in `XX XX XX ...` notation which would cause issues if copied directly from the browser.

This change will allow it to be provided in the following formats `XX XX XX ...`, `xxxxxx...` or `XX:XX:XX...`  which should cover all possible formats when retrieving the SHA256 fingerprint from a certificate on the firewall.